requirements.txt
# Utility Packages
requests
python-dotenv
asyncio

# Auth and Security Packages
msal
jwt

# Streamlit Packages
streamlit

